Fixed loading bug, added volume support

+++ include/Player.h
@@ -25,6 +25,6 @@ void MOD_Player_play(MOD_Player* player, MOD* mod);
 
 MOD_Player_Channel* MOD_Player_Channel_create();
 void MOD_Player_Channel_free(MOD_Player_Channel* channel);
-uint8_t MOD_Player_Channel_step(MOD_Player_Channel* player_channel, MOD_Player* player, MOD* mod, MOD_Channel* channel);
+double MOD_Player_Channel_step(MOD_Player_Channel* player_channel, MOD_Player* player, MOD* mod, MOD_Channel* channel);
 
 #endif
